refactor(desktop): rename all PascalCase .tsx files to kebab-case
Renames 23 component files from PascalCase to kebab-case to match the
convention already used by all .ts files in the codebase. Updates all
import paths accordingly.
