+++ desktop/src/renderer/components/editor.tsx
