+++ desktop/src/renderer/components/footer-portal.tsx
