+++ desktop/src/renderer/components/footer.tsx
