+++ desktop/src/renderer/notes/components/editor-tabs.tsx
