+++ desktop/src/renderer/notes/components/meta-panel.tsx
