+++ desktop/src/renderer/notes/components/quick-add.tsx
