+++ desktop/src/renderer/notes/editor/note-editor.tsx
