+++ desktop/src/renderer/timeline/render/now-marker.tsx
