+++ desktop/src/renderer/timeline/render/session-bands.tsx
