+++ desktop/src/renderer/views/campaigns/campaign-manager.tsx
