+++ desktop/src/renderer/views/notes/sidebar.tsx
