+++ desktop/src/renderer/views/setup/directory-picker.tsx
